-added RepeatStart functionality
git-svn-id: https://svn.cryptool.org/CrypTool2/trunk@1814 9b0d660e-8ab2-dd11-ad6d-000c29d71a56

+++ CrypPlugins/WorkspaceManager/View/Container/PluginContainerView.xaml.cs
@@ -398,7 +398,20 @@ namespace WorkspaceManager.View.Container
 
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            this.delete();
+            Button bttn = sender as Button;
+            if (bttn.Name == "play")
+            {
+                if (model.RepeatStart)
+                    model.RepeatStart = false;
+                else
+                    model.RepeatStart = true;
+                return;
+            }
+            if (bttn.Name == "del")
+            {
+                this.delete();
+                return;
+            }
         }
 
         void PluginContainerView_MouseLeave(object sender, MouseEventArgs e)
